Bootstrapping
Got the game board laid out and styled. Got a lot of the plumbing I'll
need to modify the board including some functions to identify cells on
the board and move the contents of one cell to another. I'm working on
getting random spawns started for the beginning of the game then I'll
start tying in actual gameplay functions.