Replaced table with divs
Removed the table in favor of using divs. I refactored some of the JS to
work with the divs instead of the tds they had been working with but a
good amount of the JS is still targeted at tds.

I've got numbers randomly spawning and the distribution of 2s and 4s
seems good to me. The problem is that the random spawner occasionally
spawns a number in an occupied cell. This doesn't put two numbers in the
cell but it overwrites the contents of the cell. This can be seen by
constantly refreshing the page until the board spawns with only one
starting tile. It's possible this is caused by an array index problem
instead of randomly choosing the same cell but I doubt it.