Deuglifying
Some pretty big changes to styles. Got the colors for the board looking
nicer. Changed the board and cells to be sized statically instead of
dynamically. Made some minor changes to tile's line height to
accommodate the new size of cells and tiles. Removed the left and top
positioning for cells since their floating and margins take care of
positioning.

Still need to tune up the calculations that place tiles on the board but
that should be pretty straightforward since board sizing is static now.